requirements.txt
pillow~=5.3
adafruit_ssd1306~=1.6
pygame~=1.9
